requirements.txt
getch==1.0
holidays==0.66
joblib==1.4.2
lightgbm==4.5.0
matplotlib==3.10.1
numexpr==2.10.1
numpy==2.2.4
openmeteo_requests==1.3.0
optuna==4.2.1
pandas==2.2.3
plotly==5.24.1
PyYAML==6.0.2
PyYAML==6.0.2
Requests==2.32.3
requests_cache==0.4.13
retry_requests==2.0.0
ruamel.base==1.0.0
scikit_learn==1.6.1
scipy==1.15.2
seaborn==0.13.2
shap==0.46.0
statsmodels==0.14.4
torch==2.6.0
tqdm==4.67.1
